Avoid calling SaveHelper with size_t (MK)
git-svn-id: https://svn.code.sf.net/p/ourorgan/svn/trunk@600 69509848-2a73-49d5-a893-c4d8c934025a

+++ src/grandorgue/GOrgueSetter.cpp
@@ -701,7 +701,7 @@ void GOrgueSetter::LoadCombination(IniFileConfig& cfg)
 
 void GOrgueSetter::Save(IniFileConfig& cfg, bool prefix)
 {
-	cfg.SaveHelper(prefix, wxT("Organ"), wxT("NumberOfFrameGenerals"), m_framegeneral.size());
+	cfg.SaveHelper(prefix, wxT("Organ"), wxT("NumberOfFrameGenerals"), (int)m_framegeneral.size());
 
 	for (unsigned j = 0; j < m_framegeneral.size(); j++)
 		m_framegeneral[j]->Save(cfg, prefix);
